% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metareg.R
\name{metareg_datasets}
\alias{metareg_datasets}
\title{List available dataset ids}
\usage{
metareg_datasets()
}
\value{
Character vector of dataset ids.
}
\description{
List available dataset ids
}
\examples{
head(metareg_datasets(), 10)
}
